refactor(ums-module): extract types to separate file
